Refactor config into 3 modules: config, configfile, defaults.
