+++ package.json
@@ -13,7 +13,7 @@
   "scripts": {
     "prepare": "npm run build",
     "build": "rm -rf lib/ client/lib/ && mkdir lib && npm run generate-json-schema && tsc && tsc -p client/ && npm run lint",
-    "generate-json-schema": "typescript-json-schema tsconfig.json ConfigFile --include src/config.ts --required --noExtraProps > config.schema.json",
+    "generate-json-schema": "typescript-json-schema tsconfig.json ConfigFile --include src/configfile.ts --required --noExtraProps > config.schema.json",
     "lint": "tslint --project . --format stylish",
     "format": "find src/ client/src/ -name \"*.ts\" | xargs clang-format --style=file -i",
     "test": "npm run build && mocha"
